fix: Follow pattern of moving init logic to store
