+++ editor.planx.uk/src/pages/FlowEditor/lib/store/team.ts
@@ -1,4 +1,7 @@
 import { GeoJSONObject } from "@turf/helpers";
+import gql from "graphql-tag";
+import { client } from "lib/graphql";
+import { getTeamFromDomain } from "routes/utils";
 import { NotifyPersonalisation, TeamSettings } from "types";
 import { TeamTheme } from "types";
 import { Team } from "types";
@@ -15,6 +18,7 @@ export interface TeamStore {
 
   setTeam: (team: Team) => void;
   getTeam: () => Team;
+  initTeamStore: (teamSlugFromURLParams?: string) => Promise<void>;
 }
 
 export const teamStore: StateCreator<TeamStore, [], [], TeamStore> = (
@@ -49,4 +53,40 @@ export const teamStore: StateCreator<TeamStore, [], [], TeamStore> = (
     notifyPersonalisation: get().notifyPersonalisation,
     boundaryBBox: get().boundaryBBox,
   }),
+
+  async initTeamStore(teamSlugFromURLParams) {
+    const slug = teamSlugFromURLParams || await getTeamFromDomain(window.location.hostname)
+    const { data } = await client.query({
+      query: gql`
+        query GetTeams($slug: String!) {
+          teams(
+            order_by: { name: asc }
+            limit: 1
+            where: { slug: { _eq: $slug } }
+          ) {
+            id
+            name
+            slug
+            flows(order_by: { updated_at: desc }) {
+              slug
+              updated_at
+              operations(limit: 1, order_by: { id: desc }) {
+                actor {
+                  first_name
+                  last_name
+                }
+              }
+            }
+          }
+        }
+      `,
+      variables: { slug },
+    });
+
+    const team = data.teams[0];
+
+    if (!team) throw new Error("Team not found");
+
+    this.setTeam(team);
+  },
 });
